+++ src/common/formats/native.ts
@@ -59,10 +59,31 @@ function inspect(entry: NativeFederationEntry) {
   return [components, componentRefs, assets] as const;
 }
 
+interface NativeFederationLoadRemoteOptions {
+  remoteName: string;
+  exposedModule: string;
+}
+
 export function createNativeFederation(injector: DependencyInjector): ContainerService {
   const loader = injector.get('loader');
   const platform = injector.get('platform');
   const esm = injector.get('esm');
+
+  loader.registerResolvers({
+    '@softarc/native-federation': () => () => ({
+      async loadRemoteModule(opts: NativeFederationLoadRemoteOptions) {
+        const { exposedModule, remoteName } = opts;
+        const scope = injector.get('scope');
+        const component = await scope.getComponent({
+          remoteName,
+          remoteType: 'esm',
+          format: 'native',
+          name: exposedModule,
+        });
+        return component.exports;
+      },
+    }),
+  });
 
   return {
     async createContainer(entry: NativeFederationEntry) {