+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Improved released package on JSR (only considering native output)
 - Added more inline documentation / type declarations
 - Added `loading-template-id` to attributes of `pi-component` and `pi-slot` (#8)
+- Added `@softarc/native-federaton` stub to the native federation format (#6)
 - Added slot ordering attributes to `pi-slot` (#7)
 
 **Note**: ⚠️ Early preview: Not suitable for production yet